feat: rewrite terminal control in Rust
