--- rust_termcontrol/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "rust_termcontrol"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-crossterm = "0.27"
-libc = "0.2"
-
-[lib]
-name = "rust_termcontrol"
-crate-type = ["staticlib", "rlib"]
